WebSockets/tests/submissions/Opera/interfaces/WebSocket/url/001.html

<!doctype html>
<title>WebSockets: getting url</title>
<script src=/resources/testharness.js></script>
<script src=/resources/testharnessreport.js></script>
<script src=../../../constants.js></script>
<div id=log></div>
<script>
async_test(function(t) {
  assert_equals((new WebSocket(SCHEME_AND_DOMAIN)).url, SCHEME_AND_DOMAIN+'/');
});
</script>
