w3c-reports/12092014/efl/runned-tests/websockets/interfaces/WebSocket/url/001-expected.txt


FAIL WebSockets: getting url assert_equals: expected "ws://localhost:49001/" but got "ws://localhost:49001"(stack: assert@http://localhost:8800/resources/testharness.js:1738:42
assert_equals@http://localhost:8800/resources/testharness.js:335:15
http://localhost:8800/websockets/interfaces/WebSocket/url/001.html:9:16
step@http://localhost:8800/resources/testharness.js:798:30
test@http://localhost:8800/resources/testharness.js:73:22
global code@http://localhost:8800/websockets/interfaces/WebSocket/url/001.html:8:5)
